+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.1.2
+
+- Added functionality to invert objects as well
+- Placed the Array and Object functionality behind a user flag
+- Added `jestExpect` to the global object as a work-around to [Issue #1](https://github.com/dentemple/jest-invert/issues/1)
+
 ## v0.1.1
 
 - Update docs